+++ modules/agar/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Added `Clipboard.pPasteUrlItems` API to paste url files as clipboard items. #TINY-10275
+- Added `Files.getFileDataAsString` API to make it easier to get file data out for testing. #TINY-10275
 
 ### Removed
 - Removed the `RealClipboard.sImportToClipboard` API. #TINY-10275